Step 3: simple runtime quotation
As a first approximation we can define a Joy quotation as just an alias
to Joy.parse. This won't be very useful in practice because such
quotation would re-parse code snippet on every invocation at runtime.

Current impelmentation also doesn't yet support any form of $unquoting
as it just unconditionally uses the first part out of the string context.
We could have implemented it first for runtime case but we'll skip this
step and move straight to compile-time.