Step 13: pattern unlifting
Similarly to lifting in expression mode we introduce typeclass-
based unlifting to cover extraction of values out of quotation.

We rely on user-annotated ascriptions to figure out which type to lookup
for. Again due to limitations of current pattern mode we need to do some
extra work to get the information we need.

To inspect subpatterns we need to pass args.head to special
c.internal.subpatterns method. It will give us a list of untyped
subpatterns which we can use to extract user-defined ascriptions.